refactor: `Web3Provider` -> `web3-provider`
